push, noti 관련 수정
